Moved the tree initialization methods to classes
