+++ tests/is_terminal_test.py
@@ -6,12 +6,7 @@ from typing import Annotated
 
 from geneticengine.core.grammar import extract_grammar
 from geneticengine.core.grammar import Grammar
-from geneticengine.core.random.sources import RandomSource
-from geneticengine.core.representations.tree.treebased import PI_Grow
-from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.representations.tree.utils import GengyList
-from geneticengine.core.representations.tree.utils import relabel_nodes
-from geneticengine.core.utils import has_arguments
 from geneticengine.core.utils import is_terminal
 from geneticengine.metahandlers.ints import IntRange
 from geneticengine.metahandlers.lists import ListSizeBetween
